requirements.txt
tensorflow==2.14.0
tensorflow-addons==0.20.0
tensorflow-hub==0.13.0
matplotlib==3.7.0
numpy==1.23.5
pillow==9.5.0
streamlit==1.21.0
